Detect (and remove) Unicode BOMs when reading SQL (#1733)
## Changes

### What does this PR do?

This PR updates the code-paths that read SQL files to detect (and remove
if present) any Unicode-BOM that specifies the file encoding.

### Relevant implementation details

This depends on upstream functionality: databrickslabs/blueprint#243.

Further to using the upstream functionality, this PR removes some code:

 - `read_file()` replaced by `read_text()` from blueprint.
- `remove_bom()` this was incorrect, and has also been replaced by
`read_text()`.
 - A dead code-path was removed.

### Caveats/things to watch out for when reviewing:

This branch depends on a development branch of blueprint; this needs to
be replaced with the released version before merging.

### Linked issues

Requires: databrickslabs/blueprint#243
Partially supersedes: #1689
Resolves: #1516

### Functionality

- modified existing command: `databricks labs lakebridge transpile`

### Tests

- added unit tests

---------

Co-authored-by: Guenia Izquierdo Delgado <guenia.izquierdo@databricks.com>