Update blueprint dependency, to ensure login URLs are accepted as `host` (#1760)
## Changes

### What does this PR do?

This PR updates the blueprint dependency to enable a workaround for
dealing with the `host` profile setting when it's a login URL (instead
of just the root) for the workspace.

### Relevant implementation details

The workaround is upstream on databrickslabs/blueprint#250.

### Caveats/things to watch out for when reviewing:

Until the upstream changes are released, this branch temporarily depends
on the upstream branch that contains the fix. This allows for manual
testing with:
```sh
databricks labs install lakebridge@update-minimum-blueprint
```

### Linked issues

Depends on databrickslabs/blueprint#250.
Resolves #1706.

### Functionality

- Modifies the way the workspace client for all subcommands is
initialised.

### Tests

- manually tested