Improve LSP client directory handling (#1785)
## Changes

### What does this PR do?

This PR factors out some common code that we use for temporarily
changing the current-working-directory into a context manager. It also
updates the initialisation sequence of the LSP client to use it;
previously the current working directory was being left changed. (This
in turn often left the environment for unit tests corrupted because this
was happening in a temporary directory that was subsequently removed.)

### Caveats/things to watch out for when reviewing:

The LSP engine had some code to restore the working directory, but it
only took effect if an error occurred: the happy flow did not restore
the working directory.

Some integration tests had defensive code in place to try to
re-initialize the working directory; this was probably in place because
those tests were otherwise flaky and failing because something _else_
had corrupted the test environment. These effectively masked the issue,
so I've removed it: I'd prefer the underlying issue be fixed.

### Linked issues

Blocks #1784 (integration tests are flaky because the working directory
sometimes doesn't exist)

### Functionality

- modified existing command: `databricks labs lakebridge transpile`

### Tests

- added unit tests
- updated integration tests

---------

Co-authored-by: SundarShankar89 <72757199+sundarshankar89@users.noreply.github.com>