Support conversion of XML files with internally-specified encoding (#1828)
## Changes

This PR updates the processing loop during conversion so that the
encoding of XML files is handled correctly during reading.

Some incidental changes include:

 - Updating integration tests for improved linting.
- Updating the integration test for BB transpiling to check that output
comes through.

### Caveats/things to watch out for when reviewing:

As created, this PR builds against a non-released version of blueprint;
before merging a blueprint release will be needed and this PR will need
a minor update to reflect the updated dependency.

### Linked issues

Depends on: databrickslabs/blueprint#254
Resolves: #1822
Closes: #1689

### Functionality

- modified existing command: `databricks labs lakebridge transpile`

### Tests

- manually tested
- updated integration tests