Fix `transpile` when relative path is given for input sources (#1838)
## Changes

This PR fixes transpiling input sources that are specified with a
relative path (eg. `./a/directory`) instead of an absolute one.

Further to this, some tests were updated:

 - Minor cleanup of some existing tests. (Type-hinting, etc.)
 - A new test just for the custom `transpile_document` LSP call.

### Relevant implementation details

The fix for the bug being addressed was fairly straightforward.

The new test just for the custom `transpile_document` LSP call uncovered
an off-by-one mistake in the stubbed LSP server. Although it was
harmless, it was also fixed.

### Linked issues

Resolves #1837 

### Functionality

- modified existing command: `databricks labs lakebridge transpile`

### Tests

- manually tested
- added unit tests