Improve end-to-end tests with non-UTF-8 sources (#1843)
## Changes
### What does this PR do?

This PR improves the end-to-end integration tests that we perform with
sources that are not UTF-8 encoded. This includes:

 - Morpheus: a SQL file encoded with UTF-16.

(A Bladebridge example is not included: it turns out this fails, and
will be addressed in a subsequent PR.)

### Caveats/things to watch out for when reviewing:

Although it looks like there are changes to `wf_m_employees_load.XML`,
there aren't really: it's a side-effect of configuring git to ensure the
file has the encoding/line/file-ending that we intend. (Git prefers to
store everything internally as UTF-8, and assumes files when checked out
also are UTF-8… this PR configures git to encode the checked-out file
properly.)

### Linked issues

Relates #63

### Tests

- updated integration tests