Fix semi-broken unit tests on `main` for `installer.py` (#1890)
## Changes

This PR updates the fixture used by the installer unit tests, fixing an
error introduced in #1861 whereby it accesses (and modifies) the user's
`~/.databricks/labs/remorph-transpilers`. As a unit test, it should do
this. Coincidentally the tests pass if there is nothing installed in
that location, but can fail if there is.

### Linked issues

Follows #1861

### Tests

- fixed unit tests