Introduce `schema_fixture_factory` (#1900)
<!-- REMOVE IRRELEVANT COMMENTS BEFORE CREATING A PULL REQUEST -->
## Changes
<!-- Summary of your changes that are easy to understand. Add
screenshots when necessary, they're helpful to illustrate the before and
after state -->
### What does this PR do?
`Schema` will be extended to introduce normalization. as the constructor
is used in a few places, the changes will be big so in order to have a
smaller PR, this isolated change can go already to main. a fixture
factory is introduced that the tests go through so we have a central
place to extend later