Ensure subprocess output is logged upon failure during `install-transpile` (#1898)
## Changes

This PR was originally extracted from #1894, and aimed to improve the
logging when installing the python packages failed. Manual testing
showed this was not necessary: pip output goes straight to the console
instead of being captured, so there was nothing to log. All that's left
over is a comment explaining a confusing situation that can arise in the
debugger.

### Tests

- manually tested