Fix `install-transpile` if a transpiler backup already exists (#1893)
## Changes

This PR:

- Fixes an issue in the wheel installer whereby it would fail if there
was already a backup directory (eg. `~/.databricks/labs/remorph-saved/`)
present when it started.
- Improves logging when installing transpilers, particularly when
directories are being moved/deleted/restored.
- Factors out the common code that was present in both the maven and
wheel installers.

### Caveats/things to watch out for when reviewing:

There's still a lot of cruft in place, in particular with respect to
what the instance properties or return values mean. That notwithstanding
I think this is an improvement.

### Linked issues

Resolves: #1891

### Functionality

- modified existing command: `databricks labs lakebridge
install-transpile`

### Tests

- existing integration tests