[TECH DEBT]: Remove unnecessary `chdir()` during `transpile` (#1912)
## Changes

This PR eliminates an unnecessary call to `chdir()` when launching a LSP
server. Aside from being unnecessary, the current working directory is a
form of global state, assuming it can lead to code smells and
concurrency issues (especially during testing).

### Caveats/things to watch out for when reviewing:

A unit test is also removed: it's an example of code that assumed a
specific working directory, and only passed cooincidentally. Rather than
update it I removed it because the functionality it was testing is
covered by other tests.

### Functionality

- modified existing command: `databricks labs lakebridge transpile`

### Tests

- existing unit tests
- existing integration tests