Split recon execute into smaller modules (#1892)
<!-- REMOVE IRRELEVANT COMMENTS BEFORE CREATING A PULL REQUEST -->
## Changes
<!-- Summary of your changes that are easy to understand. Add
screenshots when necessary, they're helpful to illustrate the before and
after state -->
### What does this PR do?
This is a cherry pick of PR
https://github.com/databrickslabs/lakebridge/pull/1886 since the
refactoring can go to main and result in a smaller PR
https://github.com/databrickslabs/lakebridge/pull/1874

some functions in recon are too big and pylint rejects additions to them
due to variables count inside function and there are a few code smells
like brain files that need to be refactored and split up in order to be
more readable and extendable. This is just a start at better modularity
and there is still more work to do in order to achiever better
readability

this is a prerequisite to the work being done for issue
https://github.com/databrickslabs/lakebridge/issues/1866 as this recon
part is being extended with normalization


### Relevant implementation details

### Caveats/things to watch out for when reviewing:

### Linked issues
<!-- DOC: Link issue with a keyword: close, closes, closed, fix, fixes,
fixed, resolve, resolves, resolved. See
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
-->

Resolves #..

### Functionality

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs lakebridge ...`
- [ ] ... +add your own

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [ ] manually tested
- [ ] added unit tests
- [ ] added integration tests