Detect during install/upgrade if transpilers are installed and potentially need upgrade (#1917)
## Changes

This PR updates the install process for Lakebridge so that it detects if
transpilers are already installed. If detected, the user is notified
that they might need to be upgraded.

This is the first step to automatically upgrading the transpilers, which
is easier once this PR is in place.

### Relevant implementation details

The Databricks CLI treats upgrading a component (like `lakebridge`) as
an install: there's no separate hook in `labs.yml` for upgrades vs
installs.

### Linked issues

Progresses #1836.

### Functionality

- modified existing commands:
   - `databricks labs install lakebridge`
   - `databricks labs upgrade lakebridge`

### Tests

- manually tested
- added unit test

#### Manual Testing

Fresh install, or upgrade without transpilers installed:
```
10:29:09     INFO [d.l.lakebridge.base_install] Successfully Setup Lakebridge Components Locally
10:29:09     INFO [d.l.lakebridge.base_install] For more information, please visit https://databrickslabs.github.io/lakebridge/
```

Install or upgrade with existing transpilers installed:
```
10:33:44  WARNING [d.l.lakebridge.base_install] Detected existing Lakebridge transpilers; run 'databricks labs lakebridge install-transpile' to upgrade them.
10:33:44     INFO [d.l.lakebridge.base_install] Successfully Setup Lakebridge Components Locally
10:33:44     INFO [d.l.lakebridge.base_install] For more information, please visit https://databrickslabs.github.io/lakebridge/
```