Add DuckDB Table Validator (#1908)
## Changes

### What does this PR do?
Sometimes a profiler run will contain incomplete or corrupt outputs,
leading to delays with the customer. This PR adds a classes for defining
table validation checks and receiving validation outcomes as a list or
Spark DataFrame outputs. These checks will be used to validate that a
profiler run was successful and an accurate TCO estimate will be made
prior to sending the information to the Databricks team.

### Relevant implementation details

### Caveats/things to watch out for when reviewing:

### Linked issues

N/A

### Functionality

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs lakebridge ...`
- [X] To be hooked into the profiler CLI when #1623 is merged.

### Tests

- [X] manually tested
- [X] added unit tests
- [ ] added integration tests