Upgrade transpilers during CLI upgrade (#1978)
## Changes

This PR updates the CLI install process so that during an
install/upgrade of Lakebridge we:

- Detect any installed transpilers and upgrade them to the latest
version;
- Run the transpiler configuration questions if the transpilers were
updated.

### Relevant implementation details

The intention here is that during an upgrade we also want to ensure the
transpilers are updated. Some limitations we face:

- There's no dedicate hook (via labs.yml) for `databricks labs upgrade`
versus `databricks labs install`; upgrades are treated as installations.
- The existing transpiler installers check versions and don't do
anything if the latest version is already installed.

Incidental changes needed for this to work and be testable:

- There is a lot of transpiler-specific code that was in the main
`install.py` module; this PR moves it into the `transpiles` module.
- The existing `TranspilerInstaller` class was renamed to
`ArtifactInstaller`: it is used for installing PyPi or Maven artefacts.
- The code for installing Bladebridge and Morpheus has been moved out of
methods on the workspace installer into its own hierarchy, based on a
new `TranspilerInstaller`.
 - Accompanying tests were moved as well.

### Linked issues

Resolves #1836

### Functionality

- modified existing commands:

  - `databricks labs install lakebridge`
  - `databricks labs upgrade lakebridge`

### Tests

- [X] manually tested
- [X] added unit tests