Document internal transpiler repository interfaces (#1999)
## Changes

The primary goal of this PR is to clarify (and document) the different
types of name a transpiler can have in the existing code-base. A
secondary goal was to tighten some of the tight hints, in particular to
limit collections to their read-only abstract interfaces.

Changes include:

- Internal documentation for existing interfaces related to working with
the transpiler repository.
 - Tighten type-hints to the read-only abstract collections.
 - Some minor code changes. (Called out inline, so they're visible.)

Out of scope: dealing with the inconsistencies and quirks of the naming
that's in use.

### Tests

- existing unit tests
- existing integration tests