Fix missing user agent extras (#1989)
<!-- REMOVE IRRELEVANT COMMENTS BEFORE CREATING A PULL REQUEST -->
## Changes
<!-- Summary of your changes that are easy to understand. Add
screenshots when necessary, they're helpful to illustrate the before and
after state -->
### What does this PR do?
Recreate `WorkspaceClient` on demand in our app to get around
`user_agent_extras` limitations

### Relevant implementation details
we use `user_agent_extras` to pass telemetry information. This is
provided by the Databricks SDK with the drawback that it gets evaluated
only when `_BaseClient` gets created and not on every http call.

In lakebridge this initialization happens only once because `blueprint`
takes care of initializing the `WorkspaceClient` at the beginning of the
command. when we change the extras later in the app. it wont take
effect. `cmd` extra works because `blueprint` sets it before
initializing `WorkspaceClient`.

This is probably a bug in sdk and should be fixed by implementing proper
handling of the user agent extras in `_BaseClient`

### Linked issues
<!-- DOC: Link issue with a keyword: close, closes, closed, fix, fixes,
fixed, resolve, resolves, resolved. See
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
-->

Resolves #..

### Functionality

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs lakebridge ...`
- [ ] ... +add your own

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [X] manually tested
- [ ] added unit tests
- [ ] added integration tests