Support non-interactive installation of transpilers (#2013)
## Changes

This PR implements `--interactive=false` for the `install-transpile`
command. There are 3 possibilities with the new argument:

 - `--interactive=true`: The user can answer the prompts, if necessary.
 - `--interactive=false`: There is no user available to answer prompts.
- `--interactive=auto`: The default, detect whether it looks like the
command is being run in an interactive environment. (This is based on
whether stdin is attached to a terminal or not.)

When non-interactive, the CLI behaviour changes:

 - If there is an existing configuration, it will always be left as-is.
- If there is no existing configuration, we log a warning that this is
the case.

Further to the changes with `install-transpile`, the configuration hooks
during `databricks labs upgrade lakebridge` also run as if
`--interactive=auto` was specified: configuration is skipped if it
doesn't look we're running with a terminal.

### Relevant implementation details

Convention for Unix-like systems is to check whether stdin is attached
to a terminal via the `isatty()` call. (This will be `False` if stdin is
a pipe or reading from a file.) Python on Windows emulates this and
detects whether it's running as a Console process.

This PR updates `WorkspaceInstaller` to have a boolean flag to indicate
whether interaction is allowed; it would be better to refactor `Prompts`
to account for non-interactive environments, but that would be a much
larger set of changes. As such that's out of scope for now.

### Linked issues

Resolves #2007.

### Functionality

- updated existing commands:

   - `databricks labs lakebridge install-transpile`
   - `databricks labs upgrade lakebridge`

### Tests

- [x] manually tested
- [X] added unit tests